+++ adapters/ag2/proxy_agentapi.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+exec "/home/saif/Downloads/Antigravity-x64/resources/bin/language_server" --cloud_code_endpoint https://daily-cloudcode-pa.googleapis.com --subclient_type ide agentapi "$@"